requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
scikit-learn==1.3.0
Pillow==10.0.0
gunicorn==21.2.0
requests==2.31.0